+++ tasks/mpi/vershinina_a_integration_the_monte_carlo_method/perf_tests/main.cpp
@@ -9,15 +9,15 @@
 
 TEST(vershinina_a_integration_the_monte_carlo_method, test_pipeline_run) {
   boost::mpi::communicator world;
-  std::vector<float> in;
-  std::vector<float> global_res(1, 0);
+  std::vector<double> in;
+  std::vector<double> global_res(1, 0);
 
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
-  float count_size_vector;
+  double count_size_vector;
 
   if (world.rank() == 0) {
     count_size_vector = 4;
-    in = std::vector<float>(count_size_vector, 1);
+    in = std::vector<double>(count_size_vector, 1);
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(in.data()));
     taskDataPar->inputs_count.emplace_back(in.size());
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
@@ -26,7 +26,7 @@ TEST(vershinina_a_integration_the_monte_carlo_method, test_pipeline_run) {
 
   auto testMpiTaskParallel =
       std::make_shared<vershinina_a_integration_the_monte_carlo_method::TestMPITaskParallel>(taskDataPar);
-  testMpiTaskParallel->p = [](float x) { return exp(sin(4 * x) + 2 * pow(x, 2)); };
+  testMpiTaskParallel->p = [](double x) { return exp(sin(4 * x) + 2 * pow(x, 2)); };
   ASSERT_EQ(testMpiTaskParallel->validation(), true);
   testMpiTaskParallel->pre_processing();
   testMpiTaskParallel->run();
@@ -49,15 +49,15 @@ TEST(vershinina_a_integration_the_monte_carlo_method, test_pipeline_run) {
 
 TEST(vershinina_a_integration_the_monte_carlo_method, test_task_run) {
   boost::mpi::communicator world;
-  std::vector<float> in;
-  std::vector<float> global_res(1, 0);
+  std::vector<double> in;
+  std::vector<double> global_res(1, 0);
 
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
-  float count_size_vector;
+  double count_size_vector;
 
   if (world.rank() == 0) {
     count_size_vector = 4;
-    in = std::vector<float>(count_size_vector, 1);
+    in = std::vector<double>(count_size_vector, 1);
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(in.data()));
     taskDataPar->inputs_count.emplace_back(in.size());
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
@@ -66,7 +66,7 @@ TEST(vershinina_a_integration_the_monte_carlo_method, test_task_run) {
 
   auto testMpiTaskParallel =
       std::make_shared<vershinina_a_integration_the_monte_carlo_method::TestMPITaskParallel>(taskDataPar);
-  testMpiTaskParallel->p = [](float x) { return exp(sin(4 * x) + 2 * pow(x, 2)); };
+  testMpiTaskParallel->p = [](double x) { return exp(sin(4 * x) + 2 * pow(x, 2)); };
   ASSERT_EQ(testMpiTaskParallel->validation(), true);
   testMpiTaskParallel->pre_processing();
   testMpiTaskParallel->run();
